print all ints

+++ divis.py
@@ -4,7 +4,7 @@ def divisors(num):
 
     for x in range(num,1, -1):
         if num % x == 0:
-            print(num / x)
+            print(int(num / x))
 
 
 divisors(10)
